+++ packages/react/src/components/builder-component.component.tsx
@@ -1113,9 +1113,12 @@ export class BuilderComponent extends React.Component<
                           <div
                             data-builder-component={this.name}
                             data-builder-content-id={fullData.id}
-                            data-builder-variation-id={
-                              fullData.testVariationId || fullData.variationId || fullData.id
-                            }
+                            {...(this.isPreviewing
+                              ? {
+                                  'data-builder-variation-id':
+                                    fullData.testVariationId || fullData.variationId || fullData.id,
+                                }
+                              : {})}
                           >
                             {!codegen && this.getCss(data) && (
                               <style
@@ -1328,12 +1331,13 @@ export class BuilderComponent extends React.Component<
         ...this.state,
         updates: ((this.state && this.state.updates) || 0) + 1,
         state: Object.assign(this.rootState, {
+          // external state first (data and locale prop) to allow overriding those values in editor (which is in this.state.state)
+          ...this.externalState,
           ...this.state.state,
           location: this.locationState,
           deviceSize: this.deviceSizeState,
           device: this.device,
           ...data.state,
-          ...this.externalState,
         }),
       };
       if (this.mounted) {