+++ packages/sdks/src/blocks/image/component-info.ts
@@ -4,7 +4,7 @@ import { serializeFn } from '../util.js';
 export const componentInfo: ComponentInfo = {
   name: 'Image',
   static: true,
-  builtIn: true,
+
   image:
     'https://firebasestorage.googleapis.com/v0/b/builder-3b0a2.appspot.com/o/images%2Fbaseline-insert_photo-24px.svg?alt=media&token=4e5d0ef4-f5e8-4e57-b3a9-38d63a9b9dc4',
   defaultStyles: {