+++ packages/sdks/src/blocks/raw-text/component-info.ts
@@ -3,7 +3,7 @@ import type { ComponentInfo } from '../../types/components';
 export const componentInfo: ComponentInfo = {
   name: 'Builder:RawText',
   hideFromInsertMenu: true,
-  builtIn: true,
+
   inputs: [
     {
       name: 'text',