+++ packages/sdks/src/blocks/section/component-info.ts
@@ -3,7 +3,7 @@ import type { ComponentInfo } from '../../types/components';
 export const componentInfo: ComponentInfo = {
   name: 'Core:Section',
   static: true,
-  builtIn: true,
+
   image:
     'https://cdn.builder.io/api/v1/image/assets%2FIsxPKMo2gPRRKeakUztj1D6uqed2%2F682efef23ace49afac61748dd305c70a',
   inputs: [