+++ packages/sdks/src/context/builder.context.lite.ts
@@ -5,6 +5,7 @@ export default createContext<BuilderContextInterface>({
   content: null,
   context: {},
   state: {},
+  setState: () => {},
   apiKey: null,
   registeredComponents: {},
   inheritedStyles: {},