+++ packages/sdks/output/vue/CHANGELOG.md
@@ -1,3 +1,12 @@
+### 0.1.6
+
+- Types: removed redundant `builtIn` field in `customComponents` prop types
+
+### 0.1.5
+
+- Feature: support mutating Builder's `state` in custom element events and custom user code.
+- Feature: Global CSS Nesting (`&` operator) support
+
 ### 0.1.4
 
 No Changes.